Combine new and export dataset modal
